Basic `Rectangle` implementation

+++ src/main.rs
@@ -3,5 +3,5 @@ include!("lib.rs");
 
 fn main() {
     ids!(TEST, FOO, BAR);
-    println!("Hello {}, {}, {}", TEST, FOO, BAR);
+    println!("Hello {}, {}, {}", TEST as isize, FOO as isize, BAR as isize);
 }
